renamed files and removed uneeded pages since showcase page exists
