+++ client/src/components/ui/splash-screen.tsx
