--- client/src/pages/hello.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Hello = () => {
-  return <div>Hello from Hello</div>;
-};
-
-export default Hello;
